crawler/inrix/requirements.txt
geopandas
pandas
psycopg2_binary
Requests
SQLAlchemy
